requirements.txt
genshin
Jinja2
python-dotenv
pytz
